BACKEND/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
email-validator==2.1.0
bcrypt==4.1.2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-multipart
psycopg2-binary
numpy
opencv-python-headless
Pillow
pytesseract